working with users

+++ src/routes/dashboard/admin/users/+page.server.ts
@@ -1,9 +1,10 @@
+import { PUBLIC_DEMO_MODE } from '$env/static/public';
 import { withAuth } from '@supabase/auth-helpers-sveltekit';
 import { supabaseClient } from '$lib/server/supabase';
 // import { redirect, error } from '@sveltejs/kit';
 
-/** @type {import('./$types').LayoutServerLoad} */
-export const load: LayoutServerLoad = withAuth((async ({ session }) => {
+/** @type {import('./$types').PageServerLoad} */
+export const load: PageServerLoad = withAuth((async ({ session }) => {
 
     // console.log(session)
     const org = session.user?.app_metadata.org
@@ -32,28 +33,34 @@ export const load: LayoutServerLoad = withAuth((async ({ session }) => {
 
 /** @type {import('./$types').Actions} */
 export const actions = {
-    // create: async (event: any) => {
-        // create : async withAuth((async ({ event, session }) => {
-        create : async (event: any) => {
-            // withAuth((async ({ session }) => {
+    create: async (event: any) => {
 
         const form_data = await event.request.formData();
         const email = form_data.get('email');
-        console.log(email)
+        // console.log(email)
         const password = form_data.get('password');
-        console.log(password)
+        // console.log(password)
 
         const organization = form_data.get('organization');
-        console.log(organization)
+        // console.log(organization)
 
         const role = form_data.get('role');
-        console.log(role)
+        // console.log(role)
+
+        if (PUBLIC_DEMO_MODE=='true'){
+            return { error: true, message: "USER CREATION DISABLED IN DEMO MODE!" }
+        }
 
         const { data, error } = await supabaseClient.auth.api.createUser({
-            email, password, app_metadata: { org:organization, role:role }
+            email, password, app_metadata: { org: organization, role: role }
         })
+        // console.log(data, error)
 
-        console.log(data, error)
+        if (error) {
+            return { error: true, message: error.message }
+        } else {
+            return { error: false, message: `User ${data.email} created succesfully!` }
+        }
     },
 
     delete: async (event: any) => {
@@ -62,8 +69,17 @@ export const actions = {
         const user = form_data.get('user');
         // console.log('DELETE USER ', user)
 
-        const { data, error } = await supabaseClient.auth.api.deleteUser(user)
+        if (PUBLIC_DEMO_MODE=='true'){
+            return { error: true, message: "USER DELETION DISABLED IN DEMO MODE!" }
+        }
 
+        const { data, error } = await supabaseClient.auth.api.deleteUser(user)
         // console.log(data, error)
+
+        if (error) {
+            return { error: true, message: error.message }
+        } else {
+            return { error: false, message: `User deleted!` }
+        }
     }
 };
